+++ README.md
@@ -11,4 +11,4 @@
 
 ### 덱
 - [x] 고유한 52장의 카드를 가지고 있어야 한다
-- [ ] 요청한 수량에 맞게 카드를 뽑을 수 있다
+- [x] 요청한 수량에 맞게 카드를 뽑을 수 있다
