+++ src/main/kotlin/blackjack/domain/Deck.kt
@@ -1,0 +1,28 @@
+package blackjack.domain
+
+class Deck {
+    private val _cards = mutableListOf<Card>()
+    val cards: List<Card> get() = _cards.toList()
+
+    init {
+        _cards.addAll(generateDeck())
+        require(cards.size == DECK_SIZE) { INVALID_DECK_SIZE_ERROR_MESSAGE }
+    }
+
+    fun draw(amount: Int): List<Card> {
+        val drawCards = cards.take(amount)
+        _cards.removeAll(drawCards)
+        return drawCards
+    }
+
+    private fun generateDeck(): List<Card> = CardPattern.entries.flatMap(::createCard).shuffled()
+
+    private fun createCard(cardPattern: CardPattern): List<Card> {
+        return CardNumber.entries.map { cardNumber -> Card.create(cardNumber, cardPattern) }
+    }
+
+    companion object {
+        private const val DECK_SIZE = 52
+        private const val INVALID_DECK_SIZE_ERROR_MESSAGE = "덱은 52장의 카드로 구성되어야 합니다."
+    }
+}
